backend/requirements.txt
fastapi
uvicorn[standard]
mistralai
python-dotenv
python-multipart
websockets
